+++ experimental/experimental.cabal
@@ -1,0 +1,34 @@
+cabal-version:       2.2
+name:                experimental
+version:             0.2.0.0
+synopsis:            experimental projects for hasktorch
+-- description:
+homepage:            https://github.com/hasktorch/hasktorch#readme
+license:             BSD-3-Clause
+author:              Austin Huang
+maintainer:          hasktorch@gmail.com
+copyright:           2019 Austin Huang
+category:            Machine Learning
+build-type:          Simple
+
+common config
+  default-language:    Haskell2010
+  ghc-options:         -fplugin GHC.TypeLits.Normalise -fplugin GHC.TypeLits.KnownNat.Solver -fplugin GHC.TypeLits.Extra.Solver -fconstraint-solver-iterations=0 -rtsopts
+  build-depends:       base >= 4.7 && < 5
+                     , hasktorch
+                     , ghc-typelits-extra
+                     , ghc-typelits-knownnat
+                     , ghc-typelits-natnormalise
+                     , mtl
+
+executable dataloader-cifar10
+  import:              config
+  hs-source-dirs:      dataloader-cifar10
+  main-is:             Main.hs
+  ghc-options:         -fno-warn-partial-type-signatures -rtsopts
+  build-depends:       bytestring >= 0.10.8
+                     , random >= 1.1
+                     , safe-exceptions
+                     , datasets
+                     , streaming
+                     , JuicyPixels
