+++ hasktorch/src/Torch/Vision.hs
@@ -1,4 +1,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE QuasiQuotes #-}
 
 module Torch.Vision where
 
@@ -7,20 +10,36 @@ import qualified Prelude as P
 
 import Control.Monad (forM_)
 
+import           Control.Exception.Safe         ( try
+                                                , SomeException(..)
+                                                , throwIO
+                                                )
+import           Control.Monad                  ( forM_
+                                                , when
+                                                )
 import           Torch.Internal.Cast
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Internal as BSI
 import qualified Torch.DType as D
 import qualified Torch.Tensor                  as D
 import qualified Torch.TensorOptions           as D
+import qualified Torch.Functional              as D
 import qualified Foreign.ForeignPtr            as F
 import qualified Foreign.Ptr                   as F
 import qualified Torch.Internal.Managed.TensorFactories as LibTorch
+import qualified Language.C.Inline as C
+import           System.IO.Unsafe
+import           Data.Int
+import           Data.Word
+import qualified Codec.Picture                 as I
+import qualified Data.Vector.Storable          as V
 
 import qualified Torch.Typed.Vision as I
 import Torch.Functional hiding (take)
 import Torch.Tensor
 import Torch.NN
+
+C.include "<stdint.h>"
 
 getLabels' :: Int -> I.MnistData -> [Int] -> Tensor
 getLabels' n mnist imageIdxs =
@@ -66,3 +85,132 @@ dispImage img = do
         img' = reshape [28, 28] img
         scaled :: [[Float]] = let (mn, mx) = (min img', max img')  
             in asValue $ (img' - mn) / (mx - mn) * paletteMax
+
+readImage :: FilePath -> IO (Either String D.Tensor)
+readImage file =
+  I.readImage file >>= \case
+    Left err -> return $ Left err
+    Right img' -> return $ Right $ fromDynImage img'
+
+fromDynImage :: I.DynamicImage -> D.Tensor
+fromDynImage image = unsafePerformIO $ case image of
+  I.ImageY8     (I.Image width height vec) -> createTensor width height 1 D.UInt8 1 vec
+  I.ImageYF     (I.Image width height vec) -> createTensor width height 1 D.Float 4 vec
+  I.ImageYA8    (I.Image width height vec) -> createTensor width height 2 D.UInt8 1 vec
+  I.ImageRGB8   (I.Image width height vec) -> createTensor width height 3 D.UInt8 1 vec
+  I.ImageRGB16  (I.Image width height vec) -> createTensor width height 3 D.Int32 2 vec
+  I.ImageRGBF   (I.Image width height vec) -> createTensor width height 3 D.Float 4 vec
+  I.ImageRGBA8  (I.Image width height vec) -> createTensor width height 4 D.UInt8 1 vec
+  I.ImageYCbCr8 (I.Image width height vec) -> createTensor width height 3 D.UInt8 1 vec
+  I.ImageCMYK8  (I.Image width height vec) -> createTensor width height 4 D.UInt8 1 vec
+  I.ImageCMYK16 (I.Image width height vec) -> createTensorU16to32 width height 4 D.Int32 vec
+  I.ImageRGBA16 (I.Image width height vec) -> createTensorU16to32 width height 4 D.Int32 vec
+  I.ImageY16    (I.Image width height vec) -> createTensorU16to32 width height 2 D.Int32 vec
+  I.ImageYA16   (I.Image width height vec) -> createTensorU16to32 width height 3 D.Int32 vec
+  I.ImageY32    (I.Image width height vec) -> createTensorU32to64 width height 1 D.Int64 vec
+  where
+    createTensor width height channel dtype dtype_size vec = do
+      t <- ((cast2 LibTorch.empty_lo) :: [Int] -> D.TensorOptions -> IO D.Tensor) [1, height, width, channel] $ D.withDType dtype D.defaultOpts
+      D.withTensor t $ \ptr1 -> do
+        let (fptr, len) = V.unsafeToForeignPtr0 vec
+            whc = width * height * channel * dtype_size
+        F.withForeignPtr fptr $ \ptr2 -> do
+          BSI.memcpy (F.castPtr ptr1) (F.castPtr ptr2) len
+          return t
+    createTensorU16to32 width height channel dtype vec = do
+      t <- ((cast2 LibTorch.empty_lo) :: [Int] -> D.TensorOptions -> IO D.Tensor) [1, height, width, channel] $ D.withDType dtype D.defaultOpts
+      D.withTensor t $ \ptr1 -> do
+        let (fptr, len) = V.unsafeToForeignPtr0 vec
+            whc = fromIntegral $ width * height * channel
+        F.withForeignPtr fptr $ \ptr2 -> do
+          let src = F.castPtr ptr2
+              dst = F.castPtr ptr1
+          [C.block| void {
+              uint16_t* src = $(uint16_t* src);
+              int32_t* dst = $(int32_t* dst);
+              for(int i=0;i<$(int whc);i++){
+                 dst[i] = src[i];
+              }
+          } |]
+          return t
+    createTensorU32to64 width height channel dtype vec = do
+      t <- ((cast2 LibTorch.empty_lo) :: [Int] -> D.TensorOptions -> IO D.Tensor) [1, height, width, channel] $ D.withDType dtype D.defaultOpts
+      D.withTensor t $ \ptr1 -> do
+        let (fptr, len) = V.unsafeToForeignPtr0 vec
+            whc = fromIntegral $ width * height * channel
+        F.withForeignPtr fptr $ \ptr2 -> do
+          let src = F.castPtr ptr2
+              dst = F.castPtr ptr1
+          [C.block| void {
+              uint32_t* src = $(uint32_t* src);
+              int64_t* dst = $(int64_t* dst);
+              for(int i=0;i<$(int whc);i++){
+                 dst[i] = src[i];
+              }
+          } |]
+          return t
+
+fromImages :: [I.Image I.PixelRGB8] -> IO D.Tensor
+fromImages imgs = do
+  let num_imgs = length imgs
+      channel = 3
+      (I.Image width height _) = head imgs
+  when (num_imgs == 0) $ do
+    throwIO $ userError "The number of images should be greater than 0."
+  t <- ((cast2 LibTorch.empty_lo) :: [Int] -> D.TensorOptions -> IO D.Tensor) [num_imgs,height,width,channel] $ D.withDType D.UInt8 D.defaultOpts
+  D.withTensor t $ \ptr1 -> do
+    forM_ (zip [0..] imgs) $ \(idx,(I.Image width' height' vec)) -> do
+      let (fptr,len) = V.unsafeToForeignPtr0 vec
+          whc = width * height * channel
+      when (len /= whc) $ do
+        throwIO $ userError "vector's length is not the same as tensor' one."
+      when (width /= width') $ do
+        throwIO $ userError "image's width is not the same as first image's one"
+      when (height /= height') $ do
+        throwIO $ userError "image's height is not the same as first image's one"
+      F.withForeignPtr fptr $ \ptr2 -> do
+        BSI.memcpy (F.plusPtr (F.castPtr ptr1) (whc*idx)) ptr2 len
+  return t
+
+writeImage :: forall p. I.Pixel p => Int -> Int -> Int -> p -> D.Tensor -> IO (I.Image p)
+writeImage width height channel pixel tensor = do
+  let img@(I.Image w h vec) = I.generateImage (\_ _ -> pixel) width height :: I.Image p
+  D.withTensor tensor $ \ptr1 -> do
+    let (fptr,len) = V.unsafeToForeignPtr0 vec
+        whc = width * height * channel
+    if (len /= whc) then
+      throwIO $ userError  $ "vector's length(" ++ show len ++ ") is not the same as tensor' one."
+    else do
+      F.withForeignPtr fptr $ \ptr2 -> do
+        BSI.memcpy (F.castPtr ptr2) (F.castPtr ptr1) len
+        return img
+
+writeBitmap :: FilePath -> D.Tensor -> IO ()
+writeBitmap file tensor = do
+  case (D.shape tensor,D.dtype tensor) of
+    ([1,height,width,1],D.UInt8) -> writeImage width height 1 (0 :: I.Pixel8) tensor >>= I.writeBitmap file
+    ([1,height,width],  D.UInt8) -> writeImage width height 1 (0 :: I.Pixel8) tensor >>= I.writeBitmap file
+    ([1,height,width,3],D.UInt8) -> writeImage width height 3 (I.PixelRGB8 0 0 0) tensor >>= I.writeBitmap file
+    ([height,width,1],D.UInt8) -> writeImage width height 1 (0 :: I.Pixel8) tensor >>= I.writeBitmap file
+    ([height,width],  D.UInt8) -> writeImage width height 1 (0 :: I.Pixel8) tensor >>= I.writeBitmap file
+    ([height,width,3],D.UInt8) -> writeImage width height 3 (I.PixelRGB8 0 0 0) tensor >>= I.writeBitmap file
+    format -> throwIO $ userError $ "Can not write a image. " ++ show format ++ " is unsupported format."
+
+writePng :: FilePath -> D.Tensor -> IO ()
+writePng file tensor = do
+  case (D.shape tensor,D.dtype tensor) of
+    ([1,height,width,1],D.UInt8) -> writeImage width height 1 (0 :: I.Pixel8) tensor >>= I.writePng file
+    ([1,height,width],  D.UInt8) -> writeImage width height 1 (0 :: I.Pixel8) tensor >>= I.writePng file
+    ([1,height,width,3],D.UInt8) -> writeImage width height 3 (I.PixelRGB8 0 0 0) tensor >>= I.writePng file
+    ([height,width,1],D.UInt8) -> writeImage width height 1 (0 :: I.Pixel8) tensor >>= I.writePng file
+    ([height,width],  D.UInt8) -> writeImage width height 1 (0 :: I.Pixel8) tensor >>= I.writePng file
+    ([height,width,3],D.UInt8) -> writeImage width height 3 (I.PixelRGB8 0 0 0) tensor >>= I.writePng file
+    format -> throwIO $ userError $ "Can not write a image. " ++ show format ++ " is unsupported format."
+
+-- [batch, height, width, channel] -> [batch, channel, height, width]
+hwc2chw :: D.Tensor -> D.Tensor
+hwc2chw input = D.permute input [0,3,1,2]
+
+-- [batch, channel, height, width] -> [batch, height, width, channel]
+chw2hwc :: D.Tensor -> D.Tensor
+chw2hwc input = D.permute input [0,2,3,1]
