+++ nix/shared.nix
@@ -71,6 +71,15 @@ let
                           ; hasktorch = haskellPackagesNew."hasktorch_${postfix}"
                           ; }
                         );
+                    "hasktorch-experimental_${postfix}" =
+                      # failOnAllWarnings
+                        (haskellPackagesOld.callCabal2nix
+                          "experimental"
+                          ../experimental
+                          { libtorch-ffi = haskellPackagesNew."libtorch-ffi_${postfix}"
+                          ; hasktorch = haskellPackagesNew."hasktorch_${postfix}"
+                          ; }
+                        );
                   };
 
                 extension =
@@ -165,6 +174,7 @@ let
     };
 
     hasktorch-examples_cudatoolkit_10_1-static = pkgsOld.haskell.lib.justStaticExecutables pkgsNew.haskell.packages."${compiler}".hasktorch-examples_cudatoolkit_10_1;
+    hasktorch-experimental_cudatoolkit_10_1-static = pkgsOld.haskell.lib.justStaticExecutables pkgsNew.haskell.packages."${compiler}".hasktorch-experimental_cudatoolkit_10_1;
 
     hasktorch-typed-transformer_cudatoolkit_10_1-image = pkgsOld.dockerTools.buildImage {
       name = "hasktorch-typed-transformer_cudatoolkit_10_1";
@@ -280,6 +290,7 @@ in
 
     inherit (pkgs-linux)
       hasktorch-examples_cudatoolkit_10_1-static
+      hasktorch-experimental_cudatoolkit_10_1-static
       hasktorch-typed-transformer_cudatoolkit_10_1-image
     ;
 
@@ -296,6 +307,9 @@ in
       hasktorch-examples_cpu
       hasktorch-examples_cudatoolkit_9_2
       hasktorch-examples_cudatoolkit_10_1
+      hasktorch-experimental_cpu
+      hasktorch-experimental_cudatoolkit_9_2
+      hasktorch-experimental_cudatoolkit_10_1
     ;
     hasktorch-docs = (
       (import ./haddock-combine.nix {
@@ -320,5 +334,8 @@ in
     shell-hasktorch-examples_cpu              = (doBenchmark base-compiler.hasktorch-examples_cpu             ).env.overrideAttrs(fixmkl);
     shell-hasktorch-examples_cudatoolkit_9_2  = (doBenchmark base-compiler.hasktorch-examples_cudatoolkit_9_2 ).env.overrideAttrs(fixmkl);
     shell-hasktorch-examples_cudatoolkit_10_1 = (doBenchmark base-compiler.hasktorch-examples_cudatoolkit_10_1).env.overrideAttrs(fixmkl);
+    shell-hasktorch-experimental_cpu              = (doBenchmark base-compiler.hasktorch-experimental_cpu             ).env.overrideAttrs(fixmkl);
+    shell-hasktorch-experimental_cudatoolkit_9_2  = (doBenchmark base-compiler.hasktorch-experimental_cudatoolkit_9_2 ).env.overrideAttrs(fixmkl);
+    shell-hasktorch-experimental_cudatoolkit_10_1 = (doBenchmark base-compiler.hasktorch-experimental_cudatoolkit_10_1).env.overrideAttrs(fixmkl);
   }
